+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Extracted the scraper configuration to a dedicated tab in the asset profile dialog of the admin control panel
 - Lifted the asset profile identifier editing restriction for `MANUAL` data sources in the asset profile details dialog of the admin control panel
 - Upgraded `angular` from version `20.2.4` to `20.3.9`
 - Upgraded `Nx` from version `21.5.1` to `22.1.3`